Stop storing the entire original log lines
Instead of storing the original log line with each Request/Response, we
now check the presence of a check_term right when we instantiate the
Request/Response from a log line and only store that fact.

Later, the include filter can simply rely on that fact. Exclude filtering will
even be moved out of Request/Response, this will be done on the line level
before parsing.